clamped spectrum value to get rid of white specs

+++ code/src/grid.cpp
@@ -10,23 +10,23 @@ namespace CGL { namespace StaticScene {
     return float(rand()) / RAND_MAX;
   }
 
-  float lerp(float x, float v0, float v1) {
+  double lerp(double x, double v0, double v1) {
     return (1 - x) * v0 + x * v1;
   }
 
-  float Grid::trilerp_density(const Vector3D& v) const {
+  double Grid::trilerp_density(const Vector3D& v) const {
     // Compute coordinates and offsets for v
     Vector3D samples = Vector3D(v.x * x - 0.5, v.y * y - 0.5, v.z * z - 0.5);
     Vector3D vi = Vector3D(floor(samples.x), floor(samples.y), floor(samples.z));
     Vector3D d = samples - vi;
 
     // Trilinearly interpolate density values to compute local density
-    float d00 = lerp(d.x, D(vi), D(vi + Vector3D(1, 0, 0)));
-    float d10 = lerp(d.x, D(vi + Vector3D(0, 1, 0)), D(vi + Vector3D(1, 1, 0)));
-    float d01 = lerp(d.x, D(vi + Vector3D(0, 0, 1)), D(vi + Vector3D(1, 0, 1)));
-    float d11 = lerp(d.x, D(vi + Vector3D(0, 1, 1)), D(vi + Vector3D(1, 1, 1)));
-    float d0 = lerp(d.y, d00, d10);
-    float d1 = lerp(d.y, d01, d11);
+    double d00 = lerp(d.x, D(vi), D(vi + Vector3D(1, 0, 0)));
+    double d10 = lerp(d.x, D(vi + Vector3D(0, 1, 0)), D(vi + Vector3D(1, 1, 0)));
+    double d01 = lerp(d.x, D(vi + Vector3D(0, 0, 1)), D(vi + Vector3D(1, 0, 1)));
+    double d11 = lerp(d.x, D(vi + Vector3D(0, 1, 1)), D(vi + Vector3D(1, 1, 1)));
+    double d0 = lerp(d.y, d00, d10);
+    double d1 = lerp(d.y, d01, d11);
     return lerp(d.z, d0, d1);
   }
 
@@ -37,7 +37,7 @@ namespace CGL { namespace StaticScene {
     BBox b = get_bbox();
     if (!b.intersect(mray, tmin, tmax))
       return Spectrum(1, 1, 1);
-    float t = tmin;
+    double t = tmin;
     while (true) {
       float random = generate_rand();
       t -= std::log(1 - random) / (max_density * sigma_t);
@@ -58,19 +58,19 @@ namespace CGL { namespace StaticScene {
     BBox b = get_bbox();
     if (!b.intersect(mray, tmin, tmax))
       return Spectrum(1, 1, 1);
-    float tr = 1;
-    float t = tmin;
+    double tr = 1;
+    double t = tmin;
     while (true) {
       float random = generate_rand();
       t -= std::log(1 - random) / (max_density * sigma_t);
       if (t >= tmax)
         break;
-      float d = trilerp_density(mray.o + mray.d * t);
-      tr *= 1 - std::max(0.0f, d / max_density);
+      double d = trilerp_density(mray.o + mray.d * t);
+      tr *= 1 - std::max(0.0, d / max_density);
 
       // when transmittance gets low, start applying Russian roulette to terminate sampling
       if (tr < .1) {
-        float a = std::max(0.05f, 1 - tr);
+        double a = std::max(0.05, 1 - tr);
         if (generate_rand() < a)
           return Spectrum();
         tr /= 1 - a;
@@ -79,21 +79,21 @@ namespace CGL { namespace StaticScene {
     return Spectrum(tr, tr, tr);
   }
 
-  float Grid::p(const Vector3D& wo, const Vector3D& wi) {
+  double Grid::p(const Vector3D& wo, const Vector3D& wi) {
     return phaseHG(dot(wo, wi));
   }
-  float Grid::sample_p(const Vector3D &wo, Vector3D *wi, const Vector2D &u) {
-    float cosTheta;
+  double Grid::sample_p(const Vector3D &wo, Vector3D *wi, const Vector2D &u) {
+    double cosTheta;
     if (std::abs(g) < 1e-3)
       cosTheta = 1 - 2 * u.x;
     else {
-      float sqrTerm = (1 - g * g) / (1 - g + 2 * g * u.x);
+      double sqrTerm = (1 - g * g) / (1 - g + 2 * g * u.x);
       cosTheta = (1 + g * g - sqrTerm * sqrTerm) / (2 * g);
     }
 
     // Compute direction _wi_ for Henyey--Greenstein sample
-    float sinTheta = std::sqrt(std::max(0.0f, 1 - cosTheta * cosTheta));
-    float phi = 2 * PI * u.y;
+    double sinTheta = std::sqrt(std::max(0.0, 1 - cosTheta * cosTheta));
+    double phi = 2 * PI * u.y;
     Vector3D v1, v2;
     if (std::abs(wo.x) > std::abs(wo.y))
       v1 = Vector3D(-wo.z, 0, wo.x) / std::sqrt(wo.x * wo.x + wo.z * wo.z);
